+++ rspec-expectations/spec/spec_helper.rb
@@ -55,8 +55,6 @@ RSpec.configure do |config|
   config.include RSpec::Support::InSubProcess
 
   config.expect_with :rspec do |expectations|
-    $default_expectation_syntax = expectations.syntax # rubocop:disable Style/GlobalVars
-    expectations.syntax = :expect
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
     expectations.strict_predicate_matchers = true
   end
@@ -65,54 +63,12 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
-  config.disable_monkey_patching!
-
   # We don't want rspec-core to look in our `lib` for failure snippets.
   # When it does that, it inevitably finds this line:
   # `RSpec::Support.notify_failure(RSpec::Expectations::ExpectationNotMetError.new message)`
   # ...which isn't very helpful. Far better for it to find the expectation
   # call site in the spec.
   config.project_source_dirs -= ["lib"]
-end
-
-RSpec.shared_context "with #should enabled", :uses_should do
-  orig_syntax = nil
-
-  before(:all) do
-    orig_syntax = RSpec::Matchers.configuration.syntax
-    RSpec::Matchers.configuration.syntax = [:expect, :should]
-  end
-
-  after(:context) do
-    RSpec::Matchers.configuration.syntax = orig_syntax
-  end
-end
-
-RSpec.shared_context "with the default expectation syntax" do
-  orig_syntax = nil
-
-  before(:context) do
-    orig_syntax = RSpec::Matchers.configuration.syntax
-    RSpec::Matchers.configuration.reset_syntaxes_to_default
-  end
-
-  after(:context) do
-    RSpec::Matchers.configuration.syntax = orig_syntax
-  end
-
-end
-
-RSpec.shared_context "with #should exclusively enabled", :uses_only_should do
-  orig_syntax = nil
-
-  before(:context) do
-    orig_syntax = RSpec::Matchers.configuration.syntax
-    RSpec::Matchers.configuration.syntax = :should
-  end
-
-  after(:context) do
-    RSpec::Matchers.configuration.syntax = orig_syntax
-  end
 end
 
 RSpec.shared_context "isolate include_chain_clauses_in_custom_matcher_descriptions" do
@@ -128,6 +84,8 @@ RSpec.shared_context "with warn_about_potential_false_positives set to false", :
 
   after(:context)  { RSpec::Expectations.configuration.warn_about_potential_false_positives = original_value }
 end
+
+RSpec.configuration.include_context "with warn_about_potential_false_positives set to false", :warn_about_potential_false_positives
 
 module MinitestIntegration
   include ::RSpec::Support::InSubProcess