+++ rspec-mocks/Changelog.md
@@ -4,6 +4,7 @@
 Breaking Changes:
 
 * Ruby < 2.3 is no longer supported. (Phil Pirozhkov, rspec/rspec-mocks#1349)
+* Remove monkey-patching `should_receive`/`stub` syntax. (Phil Pirozhkov, rspec/rspec-mocks#1365)
 
 ### 3.13.5 / 2025-05-27
 [Full Changelog](https://github.com/rspec/rspec/compare/rspec-mocks-v3.13.4...rspec-mocks-v3.13.5)